requirements.txt
streamlit
python-dotenv

langchain
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface

chromadb
sentence-transformers
huggingface-hub
pypdf
PyPDF2
tqdm

